Add support of OIDC provider autodiscovery
This commit adds autodiscovery support.
If the `OIDC_PROVIDER_URL` environment variable is set and is valid, the
returned settings are automatically merged with the package configuration.
This is a convenient way to avoid the need to configure everything
manually.
In the cases where this variable is not set or the provided URL is not
valid, then we stay in the current functioning.

Signed-off-by: Eric Villard <dev@eviweb.fr>